Implemented error tracking system. Implemented packet tracking system. Fixed GMTicket packets, since I was looking at them anyway. Changed EnterVehicle to a void instead, it wasn't being used as a bool anywhere anyway.

+++ Trunk/src/hearthstone-world/MiscHandler.cpp
@@ -774,10 +774,10 @@ void WorldSession::HandleSetSelectionOpcode( WorldPacket & recv_data )
 
 void WorldSession::HandleStandStateChangeOpcode( WorldPacket & recv_data )
 {
-	uint8 animstate;
+	uint32 animstate;
 	recv_data >> animstate;
 
-	_player->SetStandState(animstate);
+	_player->SetStandState(int8(animstate));
 }
 
 void WorldSession::HandleBugOpcode( WorldPacket & recv_data )
